Prepare Format Response Data for 6.0
