--- aspnetcore/web-api/advanced/formatting/5.0samples/WebAPI5PascalCase/WebAPI5PascalCase.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-</Project>
