--- aspnetcore/web-api/advanced/formatting/sample/Model/Author.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ResponseFormattingSample.Model
-{
-    public class Author
-    {
-        public string Name { get; set; }
-        public string Twitter { get; set; }
-    }
-}
