--- aspnetcore/web-api/advanced/formatting/sample/StartupStringOutputFormatter.cs
@@ -1,51 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Formatters;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace ResponseFormattingSample
-{
-    public class StartupStringOutputFormatter
-    {
-        public StartupStringOutputFormatter(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        #region snippet
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMvc(options =>
-            {
-                // requires using Microsoft.AspNetCore.Mvc.Formatters;
-                options.OutputFormatters.RemoveType<StringOutputFormatter>();
-                options.OutputFormatters.RemoveType<HttpNoContentOutputFormatter>();
-            });
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-        }
-        #endregion
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-
-            app.UseHttpsRedirection();
-            app.UseMvc();
-        }
-    }
-}
