+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/Model/Author.cs
