+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/Startup.cs
@@ -15,13 +15,13 @@ namespace ResponseFormattingSample
 
         public IConfiguration Configuration { get; }
 
-        #region snippet
+        // <snippet>
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllers()
                 .AddXmlSerializerFormatters();
         }
-        #endregion
+        // </snippet>
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
