+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/StartupPascalCase.cs
@@ -6,26 +6,23 @@ using Microsoft.Extensions.Hosting;
 
 namespace ResponseFormattingSample
 {
-    public class StartupSerializerOptions
+    public class StartupPascalCase
     {
-        public StartupSerializerOptions(IConfiguration configuration)
+        public StartupPascalCase(IConfiguration configuration)
         {
             Configuration = configuration;
         }
 
         public IConfiguration Configuration { get; }
 
-        #region snippet
+        // <snippet>
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers().AddJsonOptions(options =>
-            {
-                options.JsonSerializerOptions.AllowTrailingCommas = true;
-                options.JsonSerializerOptions.IgnoreNullValues = true;
-            });
-
+            services.AddControllers()
+                .AddJsonOptions(options =>
+                    options.JsonSerializerOptions.PropertyNamingPolicy = null);
         }
-        #endregion
+        // </snippet>
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
