+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/StartupRespectBrowserAcceptHeader.cs
@@ -6,24 +6,25 @@ using Microsoft.Extensions.Hosting;
 
 namespace ResponseFormattingSample
 {
-    public class StartupNewtonsoftJson
+    public class StartupRespectBrowserAcceptHeader
     {
-        public StartupNewtonsoftJson(IConfiguration configuration)
+        public StartupRespectBrowserAcceptHeader(IConfiguration configuration)
         {
             Configuration = configuration;
         }
 
         public IConfiguration Configuration { get; }
 
-        #region snippet
+        // <snippet>
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers()
-                .AddNewtonsoftJson();
+            services.AddControllers(options =>
+            {
+                options.RespectBrowserAcceptHeader = true; // false by default
+            });
         }
-        #endregion
+        // </snippet>
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
