+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/StartupSerializerOptions.cs
@@ -1,33 +1,33 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace ResponseFormattingSample
 {
-    public class StartupStringOutputFormatter
+    public class StartupSerializerOptions
     {
-        public StartupStringOutputFormatter(IConfiguration configuration)
+        public StartupSerializerOptions(IConfiguration configuration)
         {
             Configuration = configuration;
         }
 
         public IConfiguration Configuration { get; }
 
-        #region snippet
+        // <snippet>
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers(options =>
+            services.AddControllers().AddJsonOptions(options =>
             {
-                // requires using Microsoft.AspNetCore.Mvc.Formatters;
-                options.OutputFormatters.RemoveType<StringOutputFormatter>();
-                options.OutputFormatters.RemoveType<HttpNoContentOutputFormatter>();
+                options.JsonSerializerOptions.AllowTrailingCommas = true;
+                options.JsonSerializerOptions.IgnoreNullValues = true;
             });
-        }
-        #endregion
 
+        }
+        // </snippet>
+
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
