+++ aspnetcore/web-api/advanced/formatting/samples/3.x/ResponseFormattingSample/StartupStringOutputFormatter.cs
@@ -1,29 +1,32 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace ResponseFormattingSample
 {
-    public class StartupRespectBrowserAcceptHeader
+    public class StartupStringOutputFormatter
     {
-        public StartupRespectBrowserAcceptHeader(IConfiguration configuration)
+        public StartupStringOutputFormatter(IConfiguration configuration)
         {
             Configuration = configuration;
         }
 
         public IConfiguration Configuration { get; }
 
-        #region snippet
+        // <snippet>
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllers(options =>
             {
-                options.RespectBrowserAcceptHeader = true; // false by default
+                // requires using Microsoft.AspNetCore.Mvc.Formatters;
+                options.OutputFormatters.RemoveType<StringOutputFormatter>();
+                options.OutputFormatters.RemoveType<HttpNoContentOutputFormatter>();
             });
         }
-        #endregion
+        // </snippet>
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
